pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Pavel94 <zhilinsky.pa@yandex.ru>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
python-dotenv = "^1.0.1"
dj-database-url = "^2.2.0"
django = "^5.0.6"
gunicorn = "^22.0.0"
psycopg2-binary = "^2.9.9"
django-bootstrap5 = "^24.2"
django-crispy-forms = "^2.1"
flake8 = "^7.1.0"
django-extensions = "^3.2.3"
django-filter = "^24.2"
rollbar = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
